requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1.4
django-bootstrap-v5==1.0.11
django-jquery==3.1.0
django-phonenumber-field==7.0.1
django-sass-processor==1.2.2
phonenumberslite==8.13.2
Pillow==9.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
django-widget_tweaks==1.4.12